Experimental support for linear capabilities
This commit lets the programmer annotate traits with modes. An included
trait must have a mode, either given at inclusion time, or given in the
trait header (this works as a default). There are currently two modes;
`linear` whose references must be treated linearly, and `unsafe` which
works just like a regular type (no restrictions on aliasing). A value of
class type may be cast into a trait type (like in `master`) or
a (composite) capability type. All modes must match exactly. Here is a
small example program to show what is currently possible:

```
-- Foo defaults to unsafe
unsafe trait Foo<t>
  require f : t
  def foo() : void
    print "Foo!"

-- Inclusion of Bar requires specifying a mode
trait Bar

-- C can be aliased freely
passive class C<t> : Foo<t> + unsafe Bar
  f : t

-- D must be treated linearly unless upcast to Foo<int>
passive class D : Foo<int> + linear Bar
  f : int

class Main
  def main() : void
    let x = new D
        y = consume x : linear Bar + unsafe Foo<int>
        z = consume y : Foo<int>
        w = z
    in
      w.foo()
```

The checking that linearity is preserved is done using an additional
pass, implemented in the file `src/typechecker/Capturechecker.hs`. This
keeps the changes made to `Typechecker.hs` to a minimum, which is nice.
Basically, the capturechecker figures out where references are captured
and when a linear reference is safe to capture.

The linear references **should** be interoperable with all the features
currently in Encore, but testing is very welcome! It should be noted
that consumes are currently not atomic (the value is read and nullified
sequentially), but this shouldn't matter in a race-free program. Linear
streams are possible but not very useful, since both `get` and `getNext`
require you to consume the stream. For any useful patterns we would need
a way to `get` and move the stream forward in one go (maybe `consume s`
should be the way to do this?).