Support for subordinate mode
Types containing a capability type with the `subord` mode cannot be
returned from its surrounding aggregate. This is checked by controlling
the target of a method call or field access and throwing an error unless
the target is `this` or is fully encapsulated itself (i.e. is a
capability type with *only* subordinate capabilities). Traits without
manifest modes are typechecked as if they were subordinate, but they
still need to be given an explicit mode at inclusion.

A known issue is that subordinate state can be leaked via closures.