Fixed upcasting bug
Before this commit, a composite type was only checked for subsumption of
operations, meaning that `T1 + T2` could be cast to `T1 * T2`. Now an
upcast must preserve the conjunctions of the subtype (i.e. not introduce
new ones). It is still safe to forget conjunctions by casting `T1 * T2`
into `T1 + T2`.