wip: adding prune lock-free streaming of value
prune can be seen as a ParT that streams a single value.

    prune :: (Fut[Maybe[t]] -> Par[t']) -> Par[t] -> Par[t']

given its signature, we can understand that the ParT given as second
argument streams the first single value out of it to the future in the
first argument. after that, we proceed by trying to kill the remaining computations.