check if actor got a killed future
a method call on an actor sends a message to the actor. this message
contains that future that it needs to fulfil. if the future has been
marked as killed, it means that the actor does not need to run the
message associated to the future and should return immediately.