wip: half solution for prune
this solution is a bit of a hack. it sets the future mode after its
creation, which is not ideal given that it may need to re-take a
lock. it has not been throughroughly tested but seems to work for the
example, i.e. it does not crash nor deadlock although it does not
prevent running the parent's future computations, which is the next step
-> figuring out why this does not happen.