add simple test and current status
Prune has been hacked into the type system. The considerations under
which it works are:

* after sending an asynchronous message to an actor, we set the future
  as linear [^1].

* future chaining works in the same way: after chaining, if the closure
  returns a linear object, then the future is marked as linear. this
  means that it can be called [^2].

* ParTs that use `>> function` or other combinators do not know whether
  their `function` returns something linear and will create a `read`
  future [^3].

[1]: this should be done at creation time of the future
[2]: again, if would be better if the future is created directly as a linear future
[3]: ideally, the closure contains the information regarding the
returned type, i.e. if the new future can be linear. this is because a
PArT can contain futures that the combinator `>>` translates into future
chaining. therefore, creating a new future for which we need to know if
it can be aliased.