fixes killing when future comes from ParT
a future that has been created by the optimised ParT has set the future's
children computations (chained computations) to null. therefore, we need
to check if the children future lists is null before we read the killed
attribute of the future. i.e.

```
  if (current->future && current->future->killed){
    ...
  }
```

if the `current->future` is null, then it cannot have been killed. this
implies that there was a single promise that was already fulfilled by a
single item in the ParT.